+++ dev/user.clj
@@ -9,10 +9,11 @@
 (comment
   (c/create-conn)
   (c/close-conn)
-  (c/set "x" 100)
+  (c/set "x" 200)
   (c/get "x")
   (c/incr "x")
   (c/expire "x" 10)
   (c/ttl "x")
+  (c/llist "x")
   (c/close-conn)
   :rcf)